requirements.txt
chardet
opencv-python-headless>=4.8
fairlearn
faker
flask
fastapi
graphviz
gradio
gymnasium
ipython
jupyter
langchain<0.3.0
langchain-core<0.3.0
langchain-openai<0.3.0
langsmith<0.2.0
nbformat
networkx
numpy<2.0
aimedorchestra                 # your editable package (`pip install -e .`)
openai>=1.0.0
pandas
pydicom>=2.4
pydot
rdkit-pypi
requests
sdv>=1.6
sentence-transformers
stable-baselines3
streamlit
torch
torch-geometric
torchvision
transformers
typing_extensions>=4.0.0
uvicorn[standard]
